requirements.txt
streamlit==1.43.0
google-genai
python-dotenv
python-docx
openpyxl
pymupdf
bcrypt==4.1.3
APScheduler==3.10.4
pandas
psycopg2-binary
xlrd
